179 Update UserApiTest, DishApiTest and WeekdayApiTest with respect to GraphqlResponseFlattenerMiddleware
